+++ src/autocomplete/Autocomplete.tsx
@@ -1,11 +1,14 @@
 import React, { useState } from 'react';
 import { FormInput } from '../formInput';
-import { Hint, Roles } from '../common';
+import { FormSelect } from '../formSelect';
+import { ErrorMessage, Hint, Roles, useHydrated } from '../common';
 import { MultiSelectOption } from '../multiSelect/Types';
 import { ResultList } from './ResultList';
 import { Selected } from '../multiSelect/components/Selected';
 import { SelectedItem } from '../multiSelect/components/SelectedItem';
 import { debounce } from 'lodash';
+import { VisuallyHidden } from '../common/components/commonTypes';
+import { FormSelectProps } from '../formSelect/FormSelect';
 
 type autocompleteProps = {
   /**
@@ -20,6 +23,28 @@ type autocompleteProps = {
    * specify the number of character to press before the option are displayed
    */
   minCharsBeforeSearch?: number;
+  /**
+   * the messsage to display if the specified number of minimum characters (minCharsBeforeSearch)
+   *  has not yet been entered by the user
+   */
+  minCharsMessage?: string;
+  /**
+   * a predicate function for determining whether to display a promptMessage
+   *  when the search input receives focus
+   */
+  promptCondition?: () => boolean;
+  /**
+   * the message to display if the promptCondition has been met
+   */
+  promptMessage?: string;
+  /**
+   * the value for the id attribute of the prompt container
+   */
+  promptId?: string;
+  /**
+   * a CSS class for styling the prompt
+   */
+  promptClassName?: string;
   /**
    * will allow to delay the filter results (the value is in ms)
    */
@@ -41,18 +66,6 @@ type autocompleteProps = {
    */
   hintClass?: string;
   /**
-   * if you want to pass extra properties
-   */
-  props?: any;
-  /**
-   * if you want to style your input passing extra properties (e.g. search icon etc)
-   */
-  suffix?: any;
-  /**
-   * if you want to style your input passing extra properties (e.g. search icon etc)
-   */
-  prefix?: any;
-  /**
    * if you want to pass an id to the result UL list
    */
   resultId?: string;
@@ -65,7 +78,7 @@ type autocompleteProps = {
    */
   resultUlStyle?: React.CSSProperties;
   /**
-   * if you want to pass a style class to the result LI list
+   * if you want to pass a style class to the result LI list (it will automatically add --odd,--even to the className to help you style the alternating rows)
    */
   resultlLiClass?: string;
   /**
@@ -101,7 +114,7 @@ type autocompleteProps = {
    */
   notFoundText?: string;
   /**
-   * event that return the selected value
+   * event that return the selected value{}
    */
   onSelected?: (value: string) => void;
   /**
@@ -120,29 +133,90 @@ type autocompleteProps = {
    * onRemoveAll of the selected listItems
    */
   onRemoveAll?: () => void;
+  /**
+   * Specifies if that field needs to be filled or not
+   */
+  required?: boolean;
+  /**
+   * allow to specify a class for the container
+   */
+  containerClassName?: string;
+  /**
+   * if a label is provided, it will be displayed
+   */
+  labelText?: string;
+  /**
+   * if a label is provided, it will provide the ability to style it
+   */
+  labelClassName?: string;
+  /**
+   * it will pass an id to the input or select element(in case of progressive enhancement)
+   */
+  id?: string;
+  /**
+   * will display an error message in three different positions (BEFORE_LABEL, BOTTOM, AFTER_LABEL and AFTER_HINT)
+   */
+  errorPosition?: AutoCompleteErrorPosition;
+  /**
+   * error message text
+   */
+  errorMessageText?: string;
+  /**
+   * error className
+   */
+  errorMessageClassName?: string;
+  /**
+   * error id
+   **/
+  errorId?: string;
+  /**
+   * visually hidden text of the error
+   */
+  errorVisuallyHiddenText?: VisuallyHidden;
+  /**
+   * if you want to pass a name for the input or for the select (in case of progressive enhancement)
+   */
+  name?: string;
+  /**
+   * it will pass extra select element(in case of progressive enhancement)
+   */
+  selectProps?: FormSelectProps;
+  /**
+   * generic parameter to pass whatever element before the input
+   **/
+  prefix?: {
+    content?: JSX.Element | string;
+    properties: React.HTMLAttributes<HTMLDivElement>;
+  };
+  /**
+   * generic parameter to pass whatever element after the input
+   **/
+  suffix?: {
+    content?: JSX.Element | string;
+    properties: React.HTMLAttributes<HTMLDivElement>;
+  };
 };
 
-//remove the default style from a button
-const unstyleBtn = {
-  background: 'none',
-  color: 'inherit',
-  border: 'none',
-  padding: 0,
-  font: 'inherit',
-  cursor: 'pointer',
-  outline: 'inherit',
-};
+export enum AutoCompleteErrorPosition {
+  BEFORE_LABEL = 'before-label',
+  BOTTOM = 'bottom',
+  AFTER_LABEL = 'after-label',
+  AFTER_HINT = 'after-hint',
+}
 
 export const Autocomplete = ({
   options,
   minCharsBeforeSearch = 1,
+  minCharsMessage = '',
+  promptCondition = () => false,
+  promptMessage = '',
+  promptId = 'input-prompt',
+  promptClassName,
   debounceMs = 0,
   inputProps,
   defaultValue = '',
   hintText,
   hintClass,
-  suffix,
-  prefix,
   multiSelect = false,
   notFoundText,
   resultId,
@@ -161,12 +235,49 @@ export const Autocomplete = ({
   onRemove,
   onRemoveAll,
   onFocus,
-  props,
+  required = false,
+  containerClassName,
+  labelText,
+  labelClassName,
+  id,
+  errorPosition,
+  errorMessageText = '',
+  errorMessageClassName,
+  errorId,
+  errorVisuallyHiddenText,
+  name,
+  selectProps,
+  prefix,
+  suffix,
 }: autocompleteProps) => {
   const [activeOption, setActiveOption] = useState<number>(0);
   const [filterList, setFilterList] = useState<string[]>([]);
   const [showOptions, setShowOptions] = useState<boolean>(false);
+  const [showPrompt, setShowPrompt] = useState<boolean>(false);
   const [userInput, setUserInput] = useState<string>(defaultValue);
+  let hydrated = useHydrated();
+
+  const showPromptMessage = (): boolean =>
+    userInput.trim().length === 0 &&
+    promptCondition() &&
+    promptMessage.length > 0;
+
+  const showMinCharsMessage = (): boolean =>
+    !showPromptMessage() &&
+    userInput.trim().length < minCharsBeforeSearch &&
+    minCharsMessage.length > 0;
+
+  const handlePrompt = () => {
+    const canShowPrompt = showMinCharsMessage() || showPromptMessage();
+
+    if (!showPrompt && canShowPrompt) {
+      setShowPrompt(true);
+    } else if (showPrompt && !canShowPrompt) {
+      setShowPrompt(false);
+    }
+  };
+
+  const hidePrompt = () => setShowPrompt(false);
 
   const delayResult = React.useMemo(
     () =>
@@ -197,6 +308,11 @@ export const Autocomplete = ({
   const debounceSearch = React.useCallback(searchMemo, [searchMemo]);
 
   const handleChange = (evt: React.FormEvent<HTMLInputElement>) => {
+    // prevent user input if promptCondition() is true
+    if (promptCondition()) {
+      return;
+    }
+
     const { value } = evt.currentTarget;
     setUserInput(value);
 
@@ -214,6 +330,10 @@ export const Autocomplete = ({
       setShowOptions(true);
     }
   }, [options, onChange]);
+
+  React.useEffect(() => {
+    handlePrompt();
+  }, [userInput]);
 
   const handleClick = (evt: React.FormEvent<HTMLInputElement>) => {
     setActiveOption(0);
@@ -243,10 +363,39 @@ export const Autocomplete = ({
   };
 
   const displayResultList = (): boolean =>
-    showOptions && userInput.length >= minCharsBeforeSearch;
+    showOptions && userInput.trim().length >= minCharsBeforeSearch;
+
+  const formInput: JSX.Element = (
+    <>
+      <FormInput
+        name={name || 'autocompleteSearch'}
+        type="text"
+        value={userInput}
+        onChange={handleChange}
+        onFocus={handlePrompt}
+        onBlur={hidePrompt}
+        prefix={prefix}
+        suffix={suffix}
+        required={required}
+        inputProps={{
+          onKeyDown,
+          autoComplete: 'off',
+          id,
+          ...inputProps,
+          ...(showPrompt && { 'aria-describedby': promptId }),
+        }}
+      />
+      {showPrompt && (
+        <div className={promptClassName} id={promptId}>
+          {showPromptMessage() && promptMessage}
+          {showMinCharsMessage() && minCharsMessage}
+        </div>
+      )}
+    </>
+  );
 
   const searchEl: JSX.Element = multiSelect ? (
-    <div className="search" style={{ ...searchContainerStyle }}>
+    <>
       <div
         role={Roles.presentation}
         style={{
@@ -254,6 +403,7 @@ export const Autocomplete = ({
           flexDirection: 'row',
           width: '100%',
           flexWrap: 'wrap',
+          position: 'relative',
         }}
       >
         {selected &&
@@ -275,18 +425,11 @@ export const Autocomplete = ({
               />
             )
           )}
-        <FormInput
-          name="autocompleteSearch"
-          type="text"
-          value={userInput}
-          onChange={handleChange}
-          inputProps={{
-            onKeyDown: onKeyDown,
-            autoComplete: 'off',
-            ...inputProps,
-          }}
-          {...props}
-        />
+        {!hydrated ? (
+          <FormSelect name="multiSelect" options={options} {...inputProps} />
+        ) : (
+          formInput
+        )}
       </div>
       <div>
         <SelectedItem
@@ -303,63 +446,82 @@ export const Autocomplete = ({
           tabIndex={0}
         />
       </div>
-    </div>
+    </>
   ) : (
-    <div className="search">
-      {hintText && (
-        <Hint text={hintText} className={hintClass} useLabel={true} />
+    <div style={{ position: 'relative' }}>
+      {errorPosition &&
+        errorPosition === AutoCompleteErrorPosition.BEFORE_LABEL && (
+          <ErrorMessage
+            text={errorMessageText}
+            className={errorMessageClassName}
+            id={errorId}
+            visuallyHiddenText={errorVisuallyHiddenText}
+          />
+        )}
+      {labelText && (
+        <label htmlFor={id} className={labelClassName}>
+          {labelText}
+        </label>
       )}
-      <FormInput
-        name="autocompleteSearch"
-        type="text"
-        value={userInput}
-        onChange={handleChange}
-        inputProps={{
-          onKeyDown: onKeyDown,
-          autoComplete: 'off',
-          ...inputProps,
-        }}
-        suffix={{
-          content: (
-            <button type="submit" style={unstyleBtn}>
-              {suffix}
-            </button>
-          ),
-        }}
-        prefix={{
-          content: (
-            <button type="submit" style={unstyleBtn}>
-              {prefix}
-            </button>
-          ),
-        }}
-        {...props}
-      />
+      {errorPosition &&
+        errorPosition === AutoCompleteErrorPosition.AFTER_LABEL && (
+          <ErrorMessage
+            text={errorMessageText}
+            className={errorMessageClassName}
+            id={errorId}
+            visuallyHiddenText={errorVisuallyHiddenText}
+          />
+        )}
+      {hintText && (
+        <Hint text={hintText} className={hintClass} useLabel={false} />
+      )}
+      {errorPosition &&
+        errorPosition === AutoCompleteErrorPosition.AFTER_HINT && (
+          <ErrorMessage
+            text={errorMessageText}
+            className={errorMessageClassName}
+            id={errorId}
+            visuallyHiddenText={errorVisuallyHiddenText}
+          />
+        )}
+      {!hydrated ? (
+        <FormSelect
+          name={name || 'select'}
+          options={options}
+          id={id}
+          defaultValue={defaultValue}
+          {...selectProps}
+        />
+      ) : (
+        formInput
+      )}
     </div>
   );
 
   return (
     <>
       {multiSelect && hintText && (
-        <Hint text={hintText} className={hintClass} useLabel={true} />
+        <Hint text={hintText} className={hintClass} useLabel={false} />
       )}
-      {searchEl}
-      {displayResultList() && (
-        <ResultList
-          list={filterList}
-          userInput={userInput}
-          activeOption={activeOption}
-          noElFoundText={notFoundText}
-          onClick={handleClick}
-          activeClass={resultActiveClass}
-          ulContainerId={resultId}
-          ulContainerClass={resultUlClass}
-          ulContainerStyle={resultUlStyle}
-          liContainerClass={resultlLiClass}
-          liContainerStyle={resultLiStyle}
-          noOptionClass={resultNoOptionClass}
-        />
-      )}
+      <div className={containerClassName} style={{ ...searchContainerStyle }}>
+        {searchEl}
+        {displayResultList() && (
+          <ResultList
+            list={filterList}
+            userInput={userInput}
+            activeOption={activeOption}
+            noElFoundText={notFoundText}
+            onClick={handleClick}
+            activeClass={resultActiveClass}
+            ulContainerId={resultId}
+            ulContainerClass={resultUlClass}
+            ulContainerStyle={resultUlStyle}
+            liContainerClass={resultlLiClass}
+            liContainerStyle={resultLiStyle}
+            noOptionClass={resultNoOptionClass}
+          />
+        )}
+      </div>
     </>
   );
 };