+++ stories/liveEdit/ButtonLive.tsx
@@ -36,8 +36,10 @@ function ButtonDemo() {
         loadingLabel="saving..." 
         customLoadingPreImage={<></>} 
         customLoadingPostImage={<></>}
+        formAction="URL"
         className=""
         style={style}
+        name=""
     />
   )
 }